requirements.txt
pandas>=2.0.3
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
dbt-postgres>=1.8.0
dbt-core[tools]
python-dotenv>=1.0.1
apache-airflow>=2.9.3
pyarrow